Switch to resolving absolute paths at import-time

+++ brainframe/cli/main.py
@@ -16,16 +16,14 @@ from brainframe.cli import (
 
 
 def main():
-    try:
-        translations_path = frozen_utils.translations_path()
-    except FileNotFoundError:
+    if not frozen_utils.TRANSLATIONS_PATH.exists():
         # This message isn't translated because we don't have access to the translations
         # file to get the translation
         print_utils.fail(
             "This distribution is missing a translations file. Please re-install the "
             "latest version of the BrainFrame CLI and try again."
         )
-    i18n.load_path.append(str(translations_path))
+    i18n.load_path.append(str(frozen_utils.TRANSLATIONS_PATH))
 
     parser = ArgumentParser(
         description=i18n.t("portal.description"), usage=i18n.t("portal.usage")
